fix: code duplication
