+++ client-app/shared/sign-in/components/azuread-provider.vue
@@ -7,27 +7,8 @@
 </template>
 
 <script setup lang="ts">
-import { globals } from "@/core/globals";
-import { getReturnUrlValue } from "@/core/utilities";
+import { useIdentityProvider } from "@/shared/sign-in/components/useIdentityProvider";
 import IdentityProvider from "./identity-provider.vue";
 
-const AUTHENTICATION_TYPE = "AzureAD";
-const DEFAULT_RETURN_URL = "/";
-
-function signIn() {
-  const origin = location.origin;
-  const returnUrl = getReturnUrlValue() ?? DEFAULT_RETURN_URL;
-
-  const oidcUrl = new URL("/signin-oidc", origin);
-  const callbackUrl = new URL("/auth/callback", origin);
-  const url = new URL("/externalsignin", origin);
-
-  callbackUrl.searchParams.set("returnUrl", returnUrl);
-  url.searchParams.set("authenticationType", AUTHENTICATION_TYPE);
-  url.searchParams.set("oidcUrl", oidcUrl.href);
-  url.searchParams.set("callbackUrl", callbackUrl.href);
-  url.searchParams.set("storeId", globals.storeId);
-
-  location.assign(url);
-}
+const { signIn } = useIdentityProvider("AzureAD", "/signin-oidc");
 </script>
